Models/requirements.txt
docopt
dpu_utils>=0.1.9
tensorflow-gpu>=1.6
numpy
